requirements.txt
streamlit>=1.27.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
openai>=1.3.5
